Remove 'FallbackUuidGenerator' and always generate Uuids ourself
This implementation is actually fine now, there's no real reason to prefer the library (and a user who does can write an IdGenerator themselves anyhow)
